requirements.txt
paddlepaddle==3.0.0rc1 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddleocr>=2.0.1
torch
torchvision
torchaudio
transformers
flask
Pillow